trocando icones do projetos e outros

+++ src/admin/delete-img.php
@@ -1,19 +1,22 @@
 <?php
+session_start();
 require_once 'init.php';
 // pega o ID da URL
 
 if (isset($_POST['delImg'])) {
-    $all_id = $_POST['delete-id'];
-    $extract_id = implode(',' , $all_id);
+    if ($_POST['delete-id'] != null) {
+        $all_id = $_POST['delete-id'];
+        $extract_id = implode(',', $all_id);
 
-$sql = "delete from img where id in($extract_id)";
-$sql_run = mysqli_query($conexao, $sql);
+        $sql = "delete from img where id in($extract_id)";
+        $sql_run = mysqli_query($conexao, $sql);
 
-if ($sql_run) {
-    header('Location: projetos.php');
+        if ($sql_run) {
+            header('Location: projetos.php');
+        }
+    } else {
+        $_SESSION['delete-error'] = true;
+        header('Location: /src/admin/projetos.php');
+        exit();
+    }
 }
-
-
-}
-
-
